Modification du type de données bancaires et du type de la liste concernant les transactions

+++ bankAccount/src/main/kotlin/com/example/bankAccount/adapters/dto/DepositMoneyRequest.kt
@@ -1,3 +1,5 @@
 package com.example.bankAccount.adapters.dto
 
-data class DepositMoneyRequest(val accountId: Long, val amount: Double)
+import java.math.BigDecimal
+
+data class DepositMoneyRequest(val accountId: Long, val amount: BigDecimal)
