+++ bankAccount/src/main/kotlin/com/example/bankAccount/domain/ports/in/DepositMoneyInput.kt
@@ -1,5 +1,7 @@
 package com.example.bankAccount.domain.ports.`in`
 
+import java.math.BigDecimal
+
 interface DepositMoneyInput {
-    fun depositMoney(accountId: Long, amount: Double)
+    fun depositMoney(accountId: Long, amount: BigDecimal)
 }
